refactor(ui): 画面レイアウトビルダを widgets/screenui へ集約する
画面足場の組み立てが states のデータ層と menuscreen に散在していた。states を
Fetch のデータ提供へ純化し、menuscreen を overlay 契約と詳細窓構築に絞るため、
画面足場の責務を新パッケージ widgets/screenui へ集約する。

- screenui を新設し TabScreen/NewTabScreen/NewPanelScreen と wrapModalRoot/centerRow を states/lib.go から移す
- LogTopY/CenterWindowRect を menuscreen から screenui へ移す。menuscreen/layout.go を削除する
- 6画面の newTabScreenUI 呼び出しを screenui.NewTabScreen に、choice の newPanelScreenUI を screenui.NewPanelScreen に寄せる
- menurt の CenterWindowRect 参照を screenui へ向ける
- layout のテストを screenui へ移す

第1段 step3。組み立て内容は不変で見た目・golden は変わらない。
