+++ internal/menurt/screen.go
@@ -14,6 +14,7 @@ import (
 	"github.com/kijimaD/ruins/internal/inputmapper"
 	"github.com/kijimaD/ruins/internal/resources"
 	"github.com/kijimaD/ruins/internal/widgets/menuscreen"
+	"github.com/kijimaD/ruins/internal/widgets/screenui"
 	w "github.com/kijimaD/ruins/internal/world"
 )
 
@@ -141,7 +142,7 @@ func (s *Screen[P]) Update(world w.World) (es.Transition[w.World], error) {
 	for i := len(s.overlays) - 1; i >= 0; i-- {
 		ov := s.overlays[i]
 		if ov.Active() {
-			if win := ov.Window(world, menuscreen.CenterWindowRect(world)); win != nil {
+			if win := ov.Window(world, screenui.CenterWindowRect(world)); win != nil {
 				s.widget.AddWindow(win)
 			}
 		}
