+++ internal/widgets/screenui/screen.go
@@ -1,0 +1,122 @@
+package screenui
+
+import (
+	"github.com/ebitenui/ebitenui"
+	"github.com/ebitenui/ebitenui/widget"
+	"github.com/kijimaD/ruins/internal/consts"
+	"github.com/kijimaD/ruins/internal/resources"
+	"github.com/kijimaD/ruins/internal/widgets/styled"
+	"github.com/kijimaD/ruins/internal/widgets/theme"
+)
+
+// TabScreen はモーダル画面の共通レイアウト入力。
+// 各画面はこの入力を渡すだけで、見出し・タブ帯・コンテンツ・フッターの配置と
+// モーダル枠、ログ回避、上詰めが標準化される。目視での位置合わせを不要にする。
+// TabLabels・Footer・Header は空なら該当行を置かず、タブの無いメニューにも使える。
+type TabScreen struct {
+	// Header は上部中央の見出し。空なら見出し行を置かない
+	Header string
+	// TabLabels はタブ帯の見出し一覧。空ならタブ帯を置かない。TabIndex を強調表示する
+	TabLabels []string
+	TabIndex  int
+	// Content は画面の中身。ページ表示行を先頭に含めると全画面で開始位置が揃う
+	Content *widget.Container
+	// Footer は下部のキー案内。空なら置かない。小さめの補助テキストで表示する
+	Footer string
+}
+
+// NewPanelScreen は中央に寄せた、内容サイズに縮む小さめパネルのメニュー画面を組む。
+// タイトル・本体・フッターを縦に積む。メインメニューやセーブロードのような簡易コマンドメニューの見た目に揃える。
+// 大きめモーダルの NewTabScreen と違い、項目数が少ない画面がエントリ数相応の大きさに収まる
+func NewPanelScreen(res resources.UIResources, title string, content *widget.Container, footer string) *ebitenui.UI {
+	panel := styled.NewVerticalContainer(
+		widget.ContainerOpts.BackgroundImage(res.Panel.Image),
+		widget.ContainerOpts.WidgetOpts(
+			widget.WidgetOpts.LayoutData(widget.AnchorLayoutData{
+				HorizontalPosition: widget.AnchorLayoutPositionCenter,
+				VerticalPosition:   widget.AnchorLayoutPositionCenter,
+			}),
+			widget.WidgetOpts.MinSize(300, 0),
+		),
+	)
+	if title != "" {
+		panel.AddChild(styled.NewMenuText(title, res))
+	}
+	panel.AddChild(content)
+	if footer != "" {
+		panel.AddChild(styled.NewDescriptionText(footer, res))
+	}
+	// 下部にログ領域ぶんの余白を確保し、その上の領域で中央寄せする。データ一覧のモーダルと
+	// 同じくログに被らないようにする
+	logReserve := consts.GameHeight - LogTopY(consts.GameHeight) + theme.Space3
+	root := widget.NewContainer(widget.ContainerOpts.Layout(
+		widget.NewAnchorLayout(widget.AnchorLayoutOpts.Padding(&widget.Insets{Bottom: logReserve})),
+	))
+	root.AddChild(panel)
+	return &ebitenui.UI{Container: root}
+}
+
+// NewTabScreen はモーダル画面の標準 UI を組み立てる。
+// 行構成は 見出し（任意）/ タブ帯（任意）/ コンテンツ / 伸縮スペーサー / フッター（任意）。
+// コンテンツは上詰めされ、フッターは下端でログの手前に収まる。呼び出し側は
+// 返り値へ詳細モーダル等のウィンドウを AddWindow できる。
+func NewTabScreen(res resources.UIResources, p TabScreen) *ebitenui.UI {
+	children := make([]widget.PreferredSizeLocateableWidget, 0, 5)
+	rowStretch := make([]bool, 0, 5)
+	add := func(c widget.PreferredSizeLocateableWidget, stretch bool) {
+		children = append(children, c)
+		rowStretch = append(rowStretch, stretch)
+	}
+
+	if p.Header != "" {
+		add(centerRow(styled.NewMenuText(p.Header, res)), false)
+	}
+	if len(p.TabLabels) > 0 {
+		add(centerRow(styled.NewTabBar(p.TabLabels, p.TabIndex, res)), false)
+	}
+	add(p.Content, false)
+	add(widget.NewContainer(), true) // 伸縮スペーサー。フッターを下端へ押す
+	if p.Footer != "" {
+		footer := styled.NewRowContainer()
+		footer.AddChild(styled.NewDescriptionText(p.Footer, res))
+		add(footer, false)
+	}
+
+	root := widget.NewContainer(
+		widget.ContainerOpts.BackgroundImage(res.Panel.Image),
+		widget.ContainerOpts.Layout(widget.NewGridLayout(
+			widget.GridLayoutOpts.Columns(1),
+			widget.GridLayoutOpts.Spacing(0, theme.Space2),
+			widget.GridLayoutOpts.Stretch([]bool{true}, rowStretch),
+			widget.GridLayoutOpts.Padding(&widget.Insets{Top: theme.Space3, Bottom: theme.Space3, Left: theme.Space3, Right: theme.Space3}),
+		)),
+	)
+	for _, c := range children {
+		root.AddChild(c)
+	}
+	return &ebitenui.UI{Container: wrapModalRoot(root)}
+}
+
+// centerRow は子を水平中央に置くアンカーコンテナを返す。タブ帯や見出しの中央寄せに使う
+func centerRow(child widget.PreferredSizeLocateableWidget) *widget.Container {
+	row := widget.NewContainer(widget.ContainerOpts.Layout(widget.NewAnchorLayout()))
+	child.GetWidget().LayoutData = widget.AnchorLayoutData{HorizontalPosition: widget.AnchorLayoutPositionCenter}
+	row.AddChild(child)
+	return row
+}
+
+// wrapModalRoot は root を画面より一回り小さい中央モーダルとして包む。
+// 外周は背景を持たず透明にし、周囲に後ろのフィールドを覗かせる。動詞タブ画面と各メニューで共通に使う。
+// 下端はゲームログの上端より上で止め、ログと重ならないようにする。
+func wrapModalRoot(root *widget.Container) *widget.Container {
+	bottom := consts.GameHeight - LogTopY(consts.GameHeight) + theme.Space3
+	outer := widget.NewContainer(
+		widget.ContainerOpts.Layout(widget.NewGridLayout(
+			widget.GridLayoutOpts.Columns(1),
+			widget.GridLayoutOpts.Stretch([]bool{true}, []bool{true}),
+			widget.GridLayoutOpts.Padding(&widget.Insets{Top: 48, Bottom: bottom, Left: 96, Right: 96}),
+		)),
+	)
+	outer.AddChild(root)
+	return outer
+}
